+++ kmesh_compile_env_pre.sh
@@ -104,6 +104,32 @@ function kmesh_set_env() {
 	export EXTRA_CFLAGS="-O0 -g"
 }
 
+function set_config() {
+	sed -i -r -e "s/($1)([ \t]*)([0-9]+)/\1\2$2/" config/kmesh_marcos_def.h
+}
+
+detect_config() {
+	local kernel_version=$(uname -r)
+
+	if [ -f "/proc/config.gz" ]; then
+		zcat /proc/config.gz 2>/dev/null
+		return $?
+	fi
+
+	if [ -f "/boot/config-$kernel_version" ]; then
+		cat "/boot/config-$kernel_version" 2>/dev/null
+		return $?
+	fi
+}
+
+CONFIG_CONTENT=$(detect_config)
+
+check_config() {
+	local config_name=$1
+	value=$(echo "$CONFIG_CONTENT" | grep -w "$config_name" | cut -d= -f2)
+	echo "$value"
+}
+
 function set_enhanced_kernel_env() {
 	# we use /usr/include/linux/bpf.h to determine the runtime environment’s
 	# support for kmesh. Considering the case of online image compilation, a
@@ -117,8 +143,13 @@ function set_enhanced_kernel_env() {
 	if [ ! -f "$KERNEL_HEADER_LINUX_BPF" ]; then
 		export KERNEL_HEADER_LINUX_BPF=/usr/include/linux/bpf.h
 	fi
+	
+	KERNEL_MAJOR=$(uname -r | awk -F '.' '{print $1}')
 
-	if grep -q "FN(parse_header_msg)" $KERNEL_HEADER_LINUX_BPF; then
+	if grep -q "FN(parse_header_msg)" $KERNEL_HEADER_LINUX_BPF ||
+	[ "$(check_config "CONFIG_DEBUG_INFO_BTF_MODULES")" == "y" ] &&
+	[ "$(check_config "CONFIG_DEBUG_INFO_BTF")" == "y" ] &&
+	[ "$KERNEL_MAJOR" -ge 6 ]; then
 		export ENHANCED_KERNEL="enhanced"
 	else
 		export ENHANCED_KERNEL="normal"
