+++ kmesh_macros_env_kernel.sh
@@ -1,34 +1,9 @@
 #!/bin/bash
 
+source ./kmesh_compile_env_pre.sh
 VERSION=$(uname -r | cut -d '.' -f 1)
 KERNEL_VERSION=$(uname -r | cut -d '-' -f 1)
 KERNEL_HEADER_LINUX_BPF=/usr/include/linux/bpf.h
-
-function set_config() {
-	sed -i -r -e "s/($1)([ \t]*)([0-9]+)/\1\2$2/" config/kmesh_marcos_def.h
-}
-
-detect_config() {
-	local kernel_version=$(uname -r)
-
-	if [ -f "/proc/config.gz" ]; then
-		zcat /proc/config.gz 2>/dev/null
-		return $?
-	fi
-
-	if [ -f "/boot/config-$kernel_version" ]; then
-		cat "/boot/config-$kernel_version" 2>/dev/null
-		return $?
-	fi
-}
-
-CONFIG_CONTENT=$(detect_config)
-
-check_config() {
-	local config_name=$1
-	value=$(echo "$CONFIG_CONTENT" | grep -E "$config_name" | cut -d= -f2)
-	echo "$value"
-}
 
 # MDA_LOOPBACK_ADDR
 if grep -q "FN(get_netns_cookie)" $KERNEL_HEADER_LINUX_BPF; then
